BE 06: Add value/type parse validation, decimal type, rename get to fetch
1. Rename AppConfig.get to .fetch to signal raise-on-miss semantics (Hash#fetch idiom)
2. Add value_matches_type validation so a value that doesn't parse as its declared type is rejected, instead of silently casting to 0/0.0
3. Add "decimal" value_type casting via BigDecimal; switch poupanca_monthly_rate to decimal to avoid float drift under compounding
4. Add specs for parse validation and decimal casting

+++ db/seeds.rb
@@ -6,7 +6,7 @@
   { key: "monte_carlo_sims", value: "1000", value_type: "integer",
     description: "Number of Monte Carlo simulations per request",
     data_source: "internal" },
-  { key: "poupanca_monthly_rate", value: "0.0067", value_type: "float",
+  { key: "poupanca_monthly_rate", value: "0.0067", value_type: "decimal",
     description: "Monthly savings account yield (poupança)",
     data_source: "BCB Selic/TR" },
   { key: "minimum_wage_cents", value: "162100", value_type: "integer",
